pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cv-parser"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "questionary>=2.0",
    "rich>=13.0",
    "openai>=1.0",
    "google-generativeai>=0.8.6",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.18"]
gemini = ["google-generativeai>=0.8", "python-docx>=1.0"]
fzf = ["pyfzf>=0.3"]
dev = ["pytest>=8.0"]

[project.scripts]
"cv-parser" = "cv_parser.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/cv_parser"]

[tool.hatch.build.targets.sdist]
include = ["src/cv_parser"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
